requirements.txt
asgiref==3.8.0
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
